+++ script/book.js
@@ -1,0 +1,6 @@
+export class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
