+++ script/elements.js
@@ -1,0 +1,4 @@
+export const inputTitle = document.getElementById('title');
+export const inputAuthor = document.getElementById('author');
+export const submitBtn = document.querySelector('.add-btn');
+export const bookSection = document.querySelector('.books');
